# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "使用go语言搭建的管理后端",
        "title": "四零专属管理后台",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/system/userRoleRel/assign": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userRoleRelController"
                ],
                "summary": "为用户分配角色",
                "operationId": "assignRolesToUser",
                "parameters": [
                    {
                        "description": "分配角色参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.UserRoleRelAssignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
        "/system/userRoleRel/listByRole": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userRoleRelController"
                ],
                "summary": "根据角色ID获取用户列表",
                "operationId": "getUsersByRoleId",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "roleId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-array_system_UserWithRoleVo"
                        }
                    }
                }
            }
        },
        "/system/userRoleRel/listByUser": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userRoleRelController"
                ],
                "summary": "根据用户ID获取角色列表",
                "operationId": "getRolesByUserId",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-array_system_UserRoleRelVo"
                        }
                    }
                }
            }
        },
        "/system/userRoleRel/remove": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userRoleRelController"
                ],
                "summary": "移除用户的角色",
                "operationId": "removeRoleFromUser",
                "parameters": [
                    {
                        "description": "移除角色参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.UserRoleRelRemoveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.ApiResult-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
...
        "response.ApiResult-array_system_UserRoleRelVo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.UserRoleRelVo"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
        "response.ApiResult-array_system_UserWithRoleVo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.UserWithRoleVo"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
...
        "system.UserRoleRelAssignRequest": {
            "type": "object",
            "required": [
                "roleIds",
                "userId"
            ],
            "properties": {
                "roleIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "system.UserRoleRelRemoveRequest": {
            "type": "object",
            "required": [
                "roleId",
                "userId"
            ],
            "properties": {
                "roleId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "system.UserRoleRelVo": {
            "type": "object",
            "properties": {
                "roleId": {
                    "type": "integer"
                },
                "roleKey": {
                    "type": "string"
                },
                "roleName": {
                    "type": "string"
                }
            }
        },
...
        "system.UserWithRoleVo": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        }
    }
}